+++ common/src/main/java/org/apache/nemo/common/punctuation/Finishmark.java
@@ -16,20 +16,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.nemo.common.ir.vertex.transform;
-
-import java.io.Serializable;
+package org.apache.nemo.common.punctuation;
 
 /**
- * Watermark event.
+ * Finish mark that notifies the data fetching is finished.
+ * This is only used for bounded source because unbounded source does not finish.
  */
-public final class Watermark implements Serializable {
-  private final long timestamp;
-  public Watermark(final long timestamp) {
-    this.timestamp = timestamp;
+public final class Finishmark {
+  private static final Finishmark INSTANCE = new Finishmark();
+
+  private Finishmark() {
+
   }
 
-  public long getTimestamp() {
-    return timestamp;
+  public static Finishmark getInstance() {
+    return INSTANCE;
   }
 }
